Move mock widgets to dedicated file
